func/requirements.txt
openai==1.42.0
azure-functions
azure-storage-blob
pypdf
python-docx
